C program silently fails on CI?
`sqrt(2)` is returning a `2.0` not e.g. `1.715...`

+++ chapter21/src/test/java/com/seaofnodes/simple/progs/newtonFloat.c
@@ -8,7 +8,7 @@ extern double test_sqrt(double);
 int main( int argc, char** argv ) {
   double epsilon = 1e-50;
   for( int i=0; i<10; i++ ) {
-    double d = test_sqrt(i), expect = sqrt(i);
+    double d = test_sqrt((double)i), expect = sqrt((double)i);
     double delta = abs(d-expect);
     printf("%d  %f   (%f)\n",i,d,delta);
     if( delta > 1e-50 )
